requirements.txt
numpy
pandas
opencv-python
matplotlib
tensorflow>=2.11
keras
ultralytics
streamlit
pillow
tqdm
scikit-learn
imgaug
onnxruntime